organized repo into dirs
