+++ requirements.txt
@@ -1,12 +1,65 @@
 affine==2.3.0
 attrs==19.2.0
+backcall==0.1.0
+bleach==3.1.0
 Click==7.0
 click-plugins==1.1.1
 cligj==0.5.0
+cycler==0.10.0
+decorator==4.4.0
+defusedxml==0.6.0
+descartes==1.1.0
+entrypoints==0.3
 Fiona==1.8.8
+importlib-metadata==0.23
+ipykernel==5.1.2
+ipython==7.8.0
+ipython-genutils==0.2.0
+ipywidgets==7.5.1
+jedi==0.15.1
+Jinja2==2.10.3
+jsonschema==3.1.1
+jupyter==1.0.0
+jupyter-client==5.3.4
+jupyter-console==6.0.0
+jupyter-core==4.6.0
+kiwisolver==1.1.0
+MarkupSafe==1.1.1
+matplotlib==3.1.1
+mistune==0.8.4
+more-itertools==7.2.0
 munch==2.3.2
+nbconvert==5.6.0
+nbformat==4.4.0
+notebook==6.0.1
 numpy==1.17.2
+pandas==0.25.1
+pandocfilters==1.4.2
+parso==0.5.1
+pexpect==4.7.0
+pickleshare==0.7.5
+prometheus-client==0.7.1
+prompt-toolkit==2.0.10
+ptyprocess==0.6.0
+Pygments==2.4.2
 pyparsing==2.4.2
+pyrsistent==0.15.4
+pyshp==2.1.0
+python-dateutil==2.8.0
+pytz==2019.3
+pyzmq==18.1.0
+qtconsole==4.5.5
 rasterio==1.1.0
+scipy==1.3.1
+seaborn==0.9.0
+Send2Trash==1.5.0
 six==1.12.0
 snuggs==1.4.7
+terminado==0.8.2
+testpath==0.4.2
+tornado==6.0.3
+traitlets==4.3.3
+wcwidth==0.1.7
+webencodings==0.5.1
+widgetsnbextension==3.5.1
+zipp==0.6.0
